Room Transitions
Added Scene Transitions using GameManager.GotoScene() method
