+++ Scenes/Entities/RoomTransition.tscn
@@ -1,0 +1,17 @@
+[gd_scene load_steps=3 format=3 uid="uid://bvh3l6wbmlgbf"]
+
+[ext_resource type="Script" path="res://Scripts/RoomTransition.cs" id="1_ax434"]
+
+[sub_resource type="RectangleShape2D" id="RectangleShape2D_qbxnu"]
+size = Vector2(8, 16)
+
+[node name="RoomTransition" type="Node2D"]
+script = ExtResource("1_ax434")
+
+[node name="Area2D" type="Area2D" parent="."]
+monitorable = false
+
+[node name="CollisionShape2D" type="CollisionShape2D" parent="Area2D"]
+shape = SubResource("RectangleShape2D_qbxnu")
+
+[connection signal="area_entered" from="Area2D" to="." method="_on_area_2d_area_entered"]
