+++ Scenes/Stages/DebugStart.tscn
@@ -1,0 +1,6 @@
+[gd_scene load_steps=2 format=3 uid="uid://c0a22fc5idmss"]
+
+[ext_resource type="Script" path="res://Scripts/DebugStart.cs" id="1_ck5jq"]
+
+[node name="DebugStart" type="Node2D"]
+script = ExtResource("1_ck5jq")
