+++ Scenes/Stages/GameStage.tscn
@@ -1,0 +1,56 @@
+[gd_scene load_steps=6 format=3 uid="uid://b85a61j58ynua"]
+
+[ext_resource type="Script" path="res://Scripts/GameStage.cs" id="1_6o28l"]
+
+[sub_resource type="RectangleShape2D" id="RectangleShape2D_k3avn"]
+size = Vector2(1032, 16)
+
+[sub_resource type="RectangleShape2D" id="RectangleShape2D_b66og"]
+size = Vector2(1032, 16)
+
+[sub_resource type="RectangleShape2D" id="RectangleShape2D_ix4nn"]
+size = Vector2(16, 144)
+
+[sub_resource type="RectangleShape2D" id="RectangleShape2D_fco8g"]
+size = Vector2(16, 144)
+
+[node name="GameStage" type="Node2D"]
+script = ExtResource("1_6o28l")
+
+[node name="BackgroundTextureRect" type="TextureRect" parent="."]
+offset_right = 40.0
+offset_bottom = 40.0
+
+[node name="TileMap" type="TileMap" parent="."]
+format = 2
+
+[node name="Camera2D" type="Camera2D" parent="."]
+process_callback = 0
+limit_left = 0
+limit_top = 0
+limit_right = 416
+limit_bottom = 128
+limit_smoothed = true
+position_smoothing_enabled = true
+drag_horizontal_enabled = true
+drag_vertical_enabled = true
+
+[node name="StageBorder_StaticBody2D" type="StaticBody2D" parent="."]
+
+[node name="CollisionShape2D_Top" type="CollisionShape2D" parent="StageBorder_StaticBody2D"]
+position = Vector2(500, -8)
+shape = SubResource("RectangleShape2D_k3avn")
+
+[node name="CollisionShape2D2_Bottom" type="CollisionShape2D" parent="StageBorder_StaticBody2D"]
+position = Vector2(500, 120)
+shape = SubResource("RectangleShape2D_b66og")
+
+[node name="CollisionShape2D3_Left" type="CollisionShape2D" parent="StageBorder_StaticBody2D"]
+position = Vector2(-8, 56)
+shape = SubResource("RectangleShape2D_ix4nn")
+
+[node name="CollisionShape2D4_Right" type="CollisionShape2D" parent="StageBorder_StaticBody2D"]
+position = Vector2(424, 56)
+shape = SubResource("RectangleShape2D_fco8g")
+
+[node name="BGMPlayer" type="AudioStreamPlayer" parent="."]
