+++ Scripts/Cave100_GameStage.cs
@@ -1,0 +1,19 @@
+using Godot;
+using System;
+
+public partial class Cave100_GameStage : GameStage
+{
+	// Called when the node enters the scene tree for the first time.
+	public override void _Ready()
+	{
+		AddSpawnPoint(GetNode<SpawnMarker>("ForestExitSpawn"));
+		
+		gameManager = GetNode<GameManager>("/root/GameManager");
+		bgmPlayer = GetNode<AudioStreamPlayer>("BGMPlayer");
+	}
+
+	// Called every frame. 'delta' is the elapsed time since the previous frame.
+	public override void _Process(double delta)
+	{
+	}
+}
