+++ project.godot
@@ -11,7 +11,7 @@ config_version=5
 [application]
 
 config/name="8px_Platformer"
-run/main_scene="res://Scenes/Stages/TestStage1.tscn"
+run/main_scene="res://Scenes/Stages/Forest100_GameStage.tscn"
 config/features=PackedStringArray("4.1", "C#", "Mobile")
 config/icon="res://icon.svg"
